requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
yfinance>=0.2.0
scipy>=1.7.0
tensorboard>=2.8.0
tqdm>=4.62.0
openai>=1.0.0
google-generativeai>=0.3.0
networkx>=2.8.0
plotly>=5.0.0